Removal of mgShape
A shape is a list of lines, the class mgShape encapsualted that concept
that then wasn't going to expand upon on it. It was complicating things,
especially where it made copies of copies of lines. All lines are
created once, and then referenced to their original creation now. Line
offsets are done when they are created by their block.

TODO: Make sure to clean up the line pointers when mgBlockElement is
deleted.