Add support for predetermined line facing for normal calculation
Most of my lines generate shapes, shape lines are defined as facing
outward from the shape. Since these are all manually put in, the line
facing as of compile time is known. Put support for predetermined line
facing in, and define known shapes ( player box and map square box )
using them. This is merely an optimization.