Add support for pre-calculated sine/cosine tables
This is an optimization intended to make the use of these functions go
from expensive to almost free. There is a loss of precision, but for
this project I consider it to be a non-issue. The ability to revert to
standard sine/cosine will be left incase this becomes an issue of
contention.