mgStressTest additions disprove mgMathFunc
Doing some real world testing on the changes made in mgMathFunc proved
them faulty. 5.6x slower, in all actuality. I'm not sure how this is
possible but regardless the startling proof in front of me both halted
and disabled progress on mgMathFunc but also proved why mgStressTest is
going to be useful in the future. Optimization is a numbers game. If you
are afraid to prove that you're right, then you're already wrong.