Remove some childish comments and start tree dump.
The binary tree will have a build option to be built with a tree dump in
it. I will probably make it a binary file in the end with a seperate
browser because the binary trees this program generates can be quite big
( 57 depth is something I've seen). I want to really, fully, evaluate
the effectiveness of a Red Black tree for this application. On that
note, there is going to be a tunable to determine at build time if this
is an unbalanced tree or a Red Black Tree.