Fixed some issues so this program builds with c4droid.
mgLinkedList made use of virtual member functions in a template class using the template member. Something that is apparently not really acceptable in C++. It compiled in VStudio and GCC here, but not on c4droid's GCC implemenetation. After some research I came to the conclusion that continuing to do so was only going to give me headaches in the future, so I worked around it.